+++ backend/controllers/order/remove.js
@@ -4,6 +4,8 @@ async function remove(req,res) {
     try{
 
         const { id } = req.params;
+        if(!Number(id) || id <= 0) return res.status(400).json({message : "Invalid Product Id Format."});
+
 
         const [ response ] = await db.query('delete from orders where user_id = ? and order_id = ?' , [req.user.userId, Number(id)])
         if(response.affectedRows === 0) return res.status(400).json({message : "Order Not Found"})
